Create files for testdata and sampledata
In order to achieve better granularity and load speed, files should be split by type, and by purpose. Files prefixed with test should be small sets of specifically made records to test with. Files prefixed with sample should be large pre-generated sets of records (which work together) in the same format.